services/ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.7.0
langchain==0.0.340
redis==5.0.1
python-multipart==0.0.6
httpx==0.25.2